Move simplifyTrack to models to prevent possible eventual re-declaration
